refactor(apiV3)!: move save_column_mappings endpoint
